disable change from user that not creator of the post

+++ backend/routes/posts.js
@@ -32,7 +32,8 @@ router.post("",
     const post = new Post({
       title: req.body.title,
       content: req.body.content,
-      imagePath: url + "/images/" + req.file.filename
+      imagePath: url + "/images/" + req.file.filename,
+      creator: req.userData.userId
     });
     post.save().then(createdPost => {
       res.status(201).json({
@@ -56,13 +57,20 @@ router.put("/:id",
     const post = new Post({
       _id: req.params.id,
       title: req.body.title,
-      content: req.body.content
+      content: req.body.content,
+      imagePath: imagePath,
+      creator: req.userData.userId
     });
 
-    Post.updateOne({ _id: req.params.id }, post).then(result => {
-      res.status(201).json({
-        message: "Post updated successfully"
-      });
+    Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post).then(result => {
+      if (result.modifiedCount > 0)
+        res.status(201).json({
+          message: "Post updated successfully"
+        });
+      else
+        res.status(401).json({
+          message: "Not authorized!"
+        });
     })
   })
 
@@ -99,11 +107,18 @@ router.get("/:id", (req, res, next) => {
 })
 
 router.delete("/:id", checkAuth, (req, res, next) => {
-  Post.deleteOne({ _id: req.params.id }).then(() => {
-    res.status(200).json({
-      message: "Post deleted successfully"
-    });
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then((result) => {
+    if (result.deletedCount > 0)
+      res.status(201).json({
+        message: "Post deleted successfully"
+      });
+    else
+      res.status(401).json({
+        message: "Not authorized!"
+      });
   })
+
+
 })
 
 module.exports = router
